requirements.txt
streamlit
google-generativeai
pandas
sentence-transformers
scikit-learn
pypdf
python-docx
beautifulsoup4
openpyxl
numpy
gspread
oauth2client
plotly
markdown
edge-tts
streamlit-agraph
networkx
google-auth
python-telegram-bot  
langdetect          
streamlit-mic-recorder
pydantic-ai
pypinyin
jieba
groq
supabase
torch
